@@ -1,4 +1,4 @@
-public enum AppConfigurationEntity: CaseIterable {
+public enum AppConfigurationEntity: CaseIterable, Sendable {
     case debug
     case qa
     case testFlight
