@@ -1,5 +1,5 @@
 import Foundation
 
-public protocol PhotosLibraryRepositoryProtocol: RepositoryProtocol {
-    func copyMediaFileToPhotos(at url: URL, completion: ((SaveMediaToPhotosErrorEntity?) -> Void)?)
+public protocol PhotosLibraryRepositoryProtocol: RepositoryProtocol, Sendable {
+    func copyMediaFileToPhotos(at url: URL) async throws(SaveMediaToPhotosErrorEntity)
 }
