@@ -1,8 +1,9 @@
+import Foundation
 import MEGADomain
 import MEGADomainMock
-import XCTest
+import Testing
 
-final class SaveNodeUseCaseTests: XCTestCase {
+struct SaveNodeUseCaseTests {
     private let photoTransfer = TransferEntity(
         path: "/mock/path/test.jpg",
         nodeHandle: 123,
@@ -77,10 +78,8 @@ final class SaveNodeUseCaseTests: XCTestCase {
         isURLVideo: Bool = false,
         movedNode: Bool = true,
         shouldFallbackToMakingOffline: Bool = true,
-        photosLibraryError: SaveMediaToPhotosErrorEntity? = nil,
-        expectationDescription: String,
-        validation: @escaping (Result<Bool, SaveMediaToPhotosErrorEntity>, MockSaveMediaToPhotoFailureHandler) -> Void
-    ) {
+        photosLibraryError: SaveMediaToPhotosErrorEntity? = nil
+    ) async -> (Bool, MockSaveMediaToPhotoFailureHandler) {
         let (sut, saveMediaToPhotoFailureHandler) = makeSUT(
             isURLImage: isURLImage,
             isURLVideo: isURLVideo,
@@ -88,164 +87,99 @@ final class SaveNodeUseCaseTests: XCTestCase {
             shouldFallbackToMakingOffline: shouldFallbackToMakingOffline,
             photosLibraryError: photosLibraryError
         )
-        let expectation = self.expectation(description: expectationDescription)
         
-        sut.saveNode(from: transfer) { result in
-            validation(result, saveMediaToPhotoFailureHandler)
-            expectation.fulfill()
-        }
-        
-        waitForExpectations(timeout: 1.0, handler: nil)
+        return (await sut.saveNode(from: transfer), saveMediaToPhotoFailureHandler)
     }
     
-    func testSaveNode_withImageURL_shouldSaveSuccessfully() {
-        executeSaveNode(
+    @Test("Saving image to Photos should succeed")
+    func testSaveNode_withImageURL_shouldSaveSuccessfully() async {
+        let (result, _) = await executeSaveNode(
             transfer: photoTransfer,
-            isURLImage: true,
-            expectationDescription: "Saving image to Photos should succeed"
-        ) { result, _ in
-            switch result {
-            case .success(let success):
-                XCTAssertTrue(success)
-            case .failure:
-                XCTFail("Expected success but got failure")
-            }
-        }
+            isURLImage: true
+        )
+        #expect(result == true)
     }
     
-    func testSaveNode_withVideoURL_shouldSaveSuccessfully() {
-        executeSaveNode(
+    @Test("Saving video to Photos should succeed")
+    func testSaveNode_withVideoURL_shouldSaveSuccessfully() async {
+        let (result, _) = await executeSaveNode(
             transfer: videoTransfer,
-            isURLVideo: true,
-            expectationDescription: "Saving video to Photos should succeed"
-        ) { result, _ in
-            switch result {
-            case .success(let success):
-                XCTAssertTrue(success)
-            case .failure:
-                XCTFail("Expected success but got failure")
-            }
-        }
+            isURLVideo: true
+        )
+        #expect(result == true)
     }
     
-    func testSaveNode_withInvalidTransfer_shouldNotSaveSuccessfully() {
-        executeSaveNode(
-            transfer: invalidTransfer,
-            expectationDescription: "Invalid transfer should return failure"
-        ) { result, _ in
-            switch result {
-            case .success(let success):
-                XCTAssertFalse(success)
-            case .failure:
-                XCTFail("Expected success but got failure")
-            }
-        }
+    @Test("Invalid transfer should return failure")
+    func testSaveNode_withInvalidTransfer_shouldNotSaveSuccessfully() async {
+        let (result, _) = await executeSaveNode(transfer: invalidTransfer)
+        #expect(result == false)
     }
     
-    func testSaveNode_whenFallbackForImage_shouldTriggerFallback() {
-        executeSaveNode(
+    @Test("Fallback to making offline for image should trigger fallback")
+    func testSaveNode_whenFallbackForImage_shouldTriggerFallback() async {
+        let (result, saveMediaToPhotoFailureHandler) = await executeSaveNode(
             transfer: photoTransfer,
             isURLImage: true,
             movedNode: false,
             shouldFallbackToMakingOffline: true,
-            photosLibraryError: .imageNotSaved,
-            expectationDescription: "Fallback to making offline for image should trigger fallback"
-        ) { result, saveMediaToPhotoFailureHandler in
-            switch result {
-            case .success(let success):
-                XCTAssertFalse(success)
-                XCTAssertEqual(saveMediaToPhotoFailureHandler.fallback_calledTimes, 1)
-            case .failure:
-                XCTFail("Expected success but got failure")
-            }
-        }
+            photosLibraryError: .imageNotSaved
+        )
+    
+        #expect(result == false)
+        #expect(saveMediaToPhotoFailureHandler.fallback_calledTimes == 1)
     }
     
-    func testSaveNode_whenFallbackForVideo_shouldTriggerFallback() {
-        executeSaveNode(
+    @Test("Fallback to making offline for video should trigger fallback")
+    func testSaveNode_whenFallbackForVideo_shouldTriggerFallback() async {
+        let (result, saveMediaToPhotoFailureHandler) = await executeSaveNode(
             transfer: videoTransfer,
             isURLVideo: true,
             movedNode: false,
             shouldFallbackToMakingOffline: true,
-            photosLibraryError: .imageNotSaved,
-            expectationDescription: "Fallback to making offline for video should trigger fallback"
-        ) { result, saveMediaToPhotoFailureHandler in
-            switch result {
-            case .success(let success):
-                XCTAssertFalse(success)
-                XCTAssertEqual(saveMediaToPhotoFailureHandler.fallback_calledTimes, 1)
-            case .failure:
-                XCTFail("Expected success but got failure")
-            }
-        }
+            photosLibraryError: .imageNotSaved
+        )
+        
+        #expect(result == false)
+        #expect(saveMediaToPhotoFailureHandler.fallback_calledTimes == 1)
     }
     
-    func testSaveNode_withNonMediaFile_shouldNotTriggerFallback() {
-        executeSaveNode(
-            transfer: nonMediaTransfer,
-            expectationDescription: "Non-media file should not trigger fallback"
-        ) { result, saveMediaToPhotoFailureHandler in
-            switch result {
-            case .success(let success):
-                XCTAssertFalse(success)
-                XCTAssertEqual(saveMediaToPhotoFailureHandler.fallback_calledTimes, 0)
-            case .failure:
-                XCTFail("Expected success but got failure")
-            }
-        }
+    @Test("Non-media file should not trigger fallback")
+    func testSaveNode_withNonMediaFile_shouldNotTriggerFallback() async {
+        let (result, saveMediaToPhotoFailureHandler) = await executeSaveNode(transfer: nonMediaTransfer)
+        #expect(result == false)
+        #expect(saveMediaToPhotoFailureHandler.fallback_calledTimes == 0)
     }
 
-    func testSaveNode_withPhotosLibraryError_shouldTriggerFallback() {
-        executeSaveNode(
+    @Test("Photos library error should trigger fallback")
+    func testSaveNode_withPhotosLibraryError_shouldTriggerFallback() async {
+        let (result, saveMediaToPhotoFailureHandler) = await executeSaveNode(
             transfer: photoTransfer,
             isURLImage: true,
             movedNode: true,
             shouldFallbackToMakingOffline: true,
-            photosLibraryError: .imageNotSaved,
-            expectationDescription: "Photos library error should trigger fallback"
-        ) { result, saveMediaToPhotoFailureHandler in
-            switch result {
-            case .success(let success):
-                XCTAssertFalse(success)
-                XCTAssertEqual(saveMediaToPhotoFailureHandler.fallback_calledTimes, 1)
-            case .failure:
-                XCTFail("Expected success but got failure")
-            }
-        }
+            photosLibraryError: .imageNotSaved
+        )
+        #expect(result == false)
+        #expect(saveMediaToPhotoFailureHandler.fallback_calledTimes == 1)
     }
 
-    func testSaveNode_withNoFallbackConfiguration_shouldNotTriggerFallback() {
-        executeSaveNode(
+    @Test("Configured not to fallback should not trigger fallback")
+    func testSaveNode_withNoFallbackConfiguration_shouldNotTriggerFallback() async {
+        let (result, saveMediaToPhotoFailureHandler) = await executeSaveNode(
             transfer: photoTransfer,
             isURLImage: true,
             movedNode: false,
-            shouldFallbackToMakingOffline: false,
-            expectationDescription: "Configured not to fallback should not trigger fallback"
-        ) { result, saveMediaToPhotoFailureHandler in
-            switch result {
-            case .success(let success):
-                XCTAssertTrue(success)
-                XCTAssertEqual(saveMediaToPhotoFailureHandler.fallback_calledTimes, 0)
-            case .failure:
-                XCTFail("Expected success but got failure")
-            }
-        }
+            shouldFallbackToMakingOffline: false
+        )
+        #expect(result == true)
+        #expect(saveMediaToPhotoFailureHandler.fallback_calledTimes == 0)
     }
     
-    func testSaveNode_withNonMediaFile_shouldCreateOfflineFile() {
+    @Test("Non-media file should create offline file")
+    func testSaveNode_withNonMediaFile_shouldCreateOfflineFile() async {
         let (sut, _) = makeSUT()
-        let expectation = self.expectation(description: "Non-media file should create offline file")
         
-        sut.saveNode(from: nonMediaTransfer) { result in
-            switch result {
-            case .success(let success):
-                XCTAssertFalse(success)
-                expectation.fulfill()
-            case .failure:
-                XCTFail("Expected success but got failure")
-            }
-        }
-        
-        waitForExpectations(timeout: 1.0, handler: nil)
+        let result = await sut.saveNode(from: nonMediaTransfer)
+        #expect(result == false)
     }
 }
