@@ -20,7 +20,7 @@ import Foundation
 /// Note:
 /// Must avoid heavy operation within mutation closure
 @propertyWrapper
-public final class Atomic<T: Sendable>: @unchecked Sendable {
+public final class Atomic<T>: @unchecked Sendable {
     
     /// The internal `NSLock` used for synchronization.
     private let lock: NSLock = NSLock()
@@ -64,5 +64,41 @@ public protocol AtomicDispatchQueueProtocol {
 extension DispatchQueue: AtomicDispatchQueueProtocol {
     public func async(flags: DispatchWorkItemFlags, execute work: @escaping @Sendable @convention(block) () -> Void) {
         async(group: nil, qos: qos, flags: flags, execute: work)
+    }
+}
+
+@propertyWrapper
+public final class WeakAtomic: @unchecked Sendable {
+    
+    /// The internal `NSLock` used for synchronization.
+    private let lock: NSLock = NSLock()
+    
+    /// The actual value being wrapped. It is weak reference
+    private weak var value: AnyObject?
+    
+    /// Returns the current instance of `WeakAtomic`, allowing access to the wrapper's methods and properties.
+    public var projectedValue: WeakAtomic {
+        return self
+    }
+    
+    /// The thread-safe wrapped value.
+    public var wrappedValue: AnyObject? {
+        lock.withLock { value }
+    }
+    
+    /// Creates an `WeakAtomic` property wrapper with an initial wrapped value
+    ///
+    /// - Parameters:
+    ///   - wrappedValue: The initial value for the wrapped property.
+    public init(wrappedValue: AnyObject?) {
+        value = wrappedValue
+    }
+    
+    /// Safely mutates the wrapped value.
+    ///
+    /// - Parameter mutation: A closure that accepts an `inout` parameter of the wrapped value type, which can be mutated safely.
+    ///  Should not perform heavy operation within the clousure, otherwise calling thread will be blocked for too long.
+    public func mutate(_ mutation: (inout AnyObject?) -> Void) {
+        lock.withLock { mutation(&self.value) }
     }
 }
