@@ -1,7 +1,7 @@
 import MEGADomain
 
 public final class MockABTestRepository: ABTestRepositoryProtocol {
-    public static var newRepo: MockABTestRepository = MockABTestRepository()
+    public static let newRepo: MockABTestRepository = MockABTestRepository()
     private let abTestValues: [ABTestFlagName: Int]
     
     public init(abTestValues: [ABTestFlagName: Int] = [:]) {
